file-searcher/requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
tqdm>=4.65.0